crdt: Include timestamp in `Change`
We can then use the real change timestamp for our COBs.
